Remove static cache
